uplift-be-flask/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0